Rename the service classes, use a type service and a version in its names
