+++ packages/crypto-kraken-exchange-services/src/services/kucoin/kucoin-rest-v1/kucoin-exchange-credentials.ts
@@ -1,4 +1,4 @@
-import { ExchangeCredentials } from '../../core';
+import { ExchangeCredentials } from '../../../core';
 
 export interface KuCoinExchangeCredentials extends ExchangeCredentials {
     apiKey: string;
