+++ packages/crypto-kraken-exchange-services/src/services/kucoin/kucoin-rest-v1/kucoin-types.ts
