+++ packages/crypto-kraken-exchange-services/src/services/yobit/yobit-rest-v3/constants.ts
@@ -1,4 +1,4 @@
-import { CurrencyPair } from '../../core';
+import { CurrencyPair } from '../../../core';
 import { YobitUtils } from './yobit-utils';
 
 const constants = {
