+++ packages/crypto-kraken-exchange-services/src/services/yobit/yobit-rest-v3/index.ts
@@ -1,0 +1,2 @@
+export { YobitRestV3 } from './yobit-service';
+export { YobitExchangeCredentials } from './yobit-exchange-credentials';
