+++ packages/crypto-kraken-exchange-services/src/services/yobit/yobit-rest-v3/yobit-response-parser.ts
@@ -1,6 +1,6 @@
 import { isArray, isNumber } from 'util';
-import { CurrencyBalance, CurrencyPair, Order, OrderBook, OrderType } from '../../core';
-import { Identified } from '../../utils/identifier';
+import { CurrencyBalance, CurrencyPair, Order, OrderBook, OrderType } from '../../../core';
+import { Identified } from '../../../utils/identifier';
 import { YobitUtils } from './yobit-utils';
 
 interface YobitTrade {
