+++ packages/crypto-kraken-exchange-services/src/services/yobit/yobit-rest-v3/yobit-utils.ts
@@ -1,4 +1,4 @@
-import { CurrencyPair, OrderType } from '../../core';
+import { CurrencyPair, OrderType } from '../../../core';
 
 export const YobitUtils = {
     getPairSymbol: (pair: CurrencyPair) => `${pair[0]}_${pair[1]}`.toLowerCase(),
